Кирей, ПИбд-21, Лаба3WEB

+++ ZakusochnayaView/FormOutput.cs
@@ -173,7 +173,7 @@ namespace ZakusochnayaView
                     {
                         Id = id.Value,
                         OutputName = textBoxName.Text,
-                        Cost = Convert.ToInt32(textBoxPrice.Text),
+                        Cost = Convert.ToDecimal(textBoxPrice.Text),
                         OutputElements = productComponentBM
                     });
                 }
@@ -182,7 +182,7 @@ namespace ZakusochnayaView
                     service.AddElement(new OutputBindingModel
                     {
                         OutputName = textBoxName.Text,
-                        Cost = Convert.ToInt32(textBoxPrice.Text),
+                        Cost = Convert.ToDecimal(textBoxPrice.Text),
                         OutputElements = productComponentBM
                     });
                 }
